Adding filter for sort comments

+++ api/controller/CommentController.php
@@ -33,14 +33,15 @@ class commentController
     }
 
     function getCommentsByProductId($params = [])
-    {
-        
+    {        
         try { 
-            if (!empty($params)) {
+            if (!empty($params) && isset($params[':ID'])) {
 
                 $comments = $this->commentModel->getCommentsByProductId($params[":ID"]);
                 $this->apiView->response($comments, 200);
             }
+            else
+                $this->apiView->response(null, 404);
         } catch (Exception $e) {
             $this->apiView->response(null, 500);
         }
@@ -74,11 +75,15 @@ class commentController
     function deleteComment($params = [])
     {
         try {
-            if (!empty($params)) {
-                $this->commentModel->delete($params[":ID"]);
-                $this->apiView->response(null, 200);
+            session_start();
+            if($_SESSION['role']==1){
+                if (!empty($params) && isset($params[":ID"])) {
+                    $this->commentModel->delete($params[":ID"]);
+                    $this->apiView->response(null, 200);
+                }else
+                    $this->apiView->response(null, 404);    
             } else
-                $this->apiView->response(null, 404);
+                $this->apiView->response(null, 401);
         } catch (Exception $e) {
             $this->apiView->response(null, 500);
         }
@@ -86,9 +91,37 @@ class commentController
     }
 
     function getFilteredComments($params = [] ){
-     
-     echo $params["SORT"];
-     die();
+
+        try{
+            if(!empty($params) && !empty($params[":ID"]) && !empty($params[":ORDER"]) && !empty($params[":SORT"] && isset($params[":SORT"]) && isset($params[":ORDER"]) && isset($params[":ID"]))){
+                $sort= '';
+                $order = '';
+                if($params[":ORDER"] == 'asc')
+                    $order = 'ASC';
+                else if($params[":ORDER"] == 'desc')
+                    $order = 'DESC';
+                else
+                    $this->apiView->response(null, 404);
+            
+                if($params[":SORT"] == 'punctuation')
+                    $sort= 'punctuation';
+                else if($params[":SORT"] == 'date')
+                    $sort= 'date';
+                else
+                    $this->apiView->response(null, 404);           
+
+                $comments = $this->commentModel->getFiltered($order, $sort, $params[":ID"]);
+                $this->apiView->response($comments, 200);          
+            }else{
+                $this->apiView->response(null, 404);
+            }
+      
+        }catch(Exception $e){
+            $this->apiView->response(null, 500);
+    }
+       
+        
+  
      
         
     }
